Patch some stuff in boot and bios to make development a bit easier
